Added difficulty enum, switch instead of if-else, code motion, reduced array references and simplified logic, high scores binary search

+++ minesweeper/Time.java
@@ -5,7 +5,7 @@ class Time {
     startTime = System.currentTimeMillis();
   }
 	
-	int timeElapsed() {
-		return (int) ((System.currentTimeMillis() - startTime) / 1000);
+	long timeElapsed() {
+		return (System.currentTimeMillis() - startTime) / 1000;
 	}
 }
